fix folder structure widget target
