+++ EngineerMemoWidgets/App/MemoWidgetBundle.swift
@@ -1,0 +1,9 @@
+import SwiftUI
+import WidgetKit
+
+@main
+struct MemoWidgetBundle: WidgetBundle {
+    var body: some Widget {
+        MemoWidget()
+    }
+}
