--- EngineerMemoWidgets/EngineerMemoWidgets.swift
@@ -1,85 +1,0 @@
-import Intents
-import SwiftUI
-import WidgetKit
-
-struct Provider: IntentTimelineProvider {
-    func placeholder(in context: Context) -> SimpleEntry {
-        .init(
-            date: .init(),
-            configuration: .init()
-        )
-    }
-
-    func getSnapshot(
-        for configuration: ConfigurationIntent,
-        in context: Context,
-        completion: @escaping (SimpleEntry) -> Void
-    ) {
-        let entry = SimpleEntry(
-            date: .init(),
-            configuration: configuration
-        )
-
-        completion(entry)
-    }
-
-    func getTimeline(
-        for configuration: ConfigurationIntent,
-        in context: Context,
-        completion: @escaping (Timeline<Entry>) -> Void
-    ) {
-        var entries: [SimpleEntry] = []
-
-        // Generate a timeline consisting of five entries an hour apart, starting from the current date.
-        let currentDate = Date()
-        for hourOffset in 0 ..< 5 {
-            let entryDate = Calendar.current.date(byAdding: .hour, value: hourOffset, to: currentDate)!
-            let entry = SimpleEntry(date: entryDate, configuration: configuration)
-            entries.append(entry)
-        }
-
-        let timeline = Timeline(entries: entries, policy: .atEnd)
-        completion(timeline)
-    }
-}
-
-struct SimpleEntry: TimelineEntry {
-    let date: Date
-    let configuration: ConfigurationIntent
-}
-
-struct EngineerMemoWidgetsEntryView: View {
-    var entry: Provider.Entry
-
-    var body: some View {
-        Text(entry.date, style: .time)
-    }
-}
-
-struct EngineerMemoWidgets: Widget {
-    let kind = "EngineerMemoWidgets"
-
-    var body: some WidgetConfiguration {
-        IntentConfiguration(
-            kind: kind,
-            intent: ConfigurationIntent.self,
-            provider: Provider()
-        ) { entry in
-            EngineerMemoWidgetsEntryView(entry: entry)
-        }
-        .configurationDisplayName("My Widget")
-        .description("This is an example widget.")
-    }
-}
-
-struct EngineerMemoWidgets_Previews: PreviewProvider {
-    static var previews: some View {
-        EngineerMemoWidgetsEntryView(
-            entry: .init(
-                date: .init(),
-                configuration: .init()
-            )
-        )
-        .previewContext(WidgetPreviewContext(family: .systemSmall))
-    }
-}
