--- EngineerMemoWidgets/EngineerMemoWidgetsBundle.swift
@@ -1,9 +1,0 @@
-import SwiftUI
-import WidgetKit
-
-@main
-struct EngineerMemoWidgetsBundle: WidgetBundle {
-    var body: some Widget {
-        EngineerMemoWidgets()
-    }
-}
