+++ EngineerMemoWidgets/Widget/Entry.swift
@@ -1,0 +1,6 @@
+import WidgetKit
+
+struct Entry: TimelineEntry {
+    let date: Date
+    let configuration: ConfigurationIntent
+}
