+++ EngineerMemoWidgets/WidgetView/EntryView.swift
@@ -1,0 +1,9 @@
+import SwiftUI
+
+struct EntryView: View {
+    var entry: Provider.Entry
+
+    var body: some View {
+        Text(entry.date, style: .time)
+    }
+}
