+++ project.yml
@@ -37,8 +37,8 @@ targets:
     type: application
     platform: iOS
     sources:
-      - path: EngineerMemo
-      - path: EngineerMemoWidgets/EngineerMemoWidgets.intentdefinition
+      - EngineerMemo
+      - EngineerMemoWidgets/EngineerMemoWidgets.intentdefinition
     settings:
       base:
         DEVELOPMENT_LANGUAGE: jp
